ref: split the CameraOpencv class in two.
One class for Linux using v4l2-ctl and the other for other situation
