+++ src/crappy/camera/opencv_camera_v4l2.py
